rename CommandsTab -> PromptsTab
