+++ vscode/webviews/CodyPanel.tsx
@@ -5,7 +5,7 @@ import type { ConfigurationSubsetForWebview, LocalEnv } from '../src/chat/protoc
 import styles from './App.module.css'
 import { Chat } from './Chat'
 import { TabContainer, TabRoot } from './components/shadcn/ui/tabs'
-import { AccountTab, CommandsTab, HistoryTab, SettingsTab, TabsBar, View } from './tabs'
+import { AccountTab, HistoryTab, PromptsTab, SettingsTab, TabsBar, View } from './tabs'
 
 /**
  * The Cody tab panel, with tabs for chat, history, prompts, etc.
@@ -33,7 +33,7 @@ export const CodyPanel: FunctionComponent<
         | 'scrollableParent'
     > &
         Pick<ComponentProps<typeof HistoryTab>, 'userHistory'> &
-        Pick<ComponentProps<typeof CommandsTab>, 'commands'>
+        Pick<ComponentProps<typeof PromptsTab>, 'commands'>
 > = ({
     view,
     setView,
@@ -84,8 +84,8 @@ export const CodyPanel: FunctionComponent<
                     />
                 )}
                 {view === View.History && <HistoryTab userHistory={userHistory} />}
-                {view === View.Commands && (
-                    <CommandsTab setView={setView} IDE={config.agentIDE} commands={commands} />
+                {view === View.Prompts && (
+                    <PromptsTab setView={setView} IDE={config.agentIDE} commands={commands} />
                 )}
                 {view === View.Account && <AccountTab userInfo={userInfo} />}
                 {view === View.Settings && <SettingsTab userInfo={userInfo} />}
